+++ packages/runtime/test/src/interpreter/opcode-list.ts
@@ -7449,7 +7449,7 @@ describe("Teal Opcodes", function () {
 			op.execute(stack);
 			assert.equal(2n, stack.pop());
 		});
-		it.only("Should return correct number of ClearStateProgramPages", function () {
+		it("Should return correct number of ClearStateProgramPages", function () {
 			interpreter.runtime.ctx.tx.apsu = Buffer.alloc(5000).fill(0);
 			const op = new Txn(["NumClearStateProgramPages"], 1, interpreter);
 			op.execute(stack);